% File src/library/multiway/man/grid-internal.Rd
% Part of the R multiway package, 
% Nathaniel E. Helwig <helwig@umn.edu>

\name{multiway-internal}

%% List of Internal Functions Called
\alias{ed2sp}
\alias{fmnls}
\alias{fsmls}
\alias{fumls}
\alias{IsplineBasis}
\alias{MsplineBasis}
\alias{parafac_3way}
\alias{parafac_4way}
\alias{parafac2_3way}
\alias{parafac2_4way}
\alias{rfunc}
\alias{tucker_3way}
\alias{tucker_4way}

\title{Internal Multi-Way Functions}
\description{
  Internal Multi-Way Functions
}
\details{
  These functions are not to be called by the user.
}
\keyword{ internal }

