% File src/library/multiway/man/grid-internal.Rd
% Part of the R multiway package, 
% Nathaniel E. Helwig <helwig@umn.edu>

\name{multiway-internal}

%% List of Internal Functions Called
\alias{cpd_nway}
\alias{cpd_nwayna}
\alias{ed2sp}
\alias{initcmls}
\alias{mcr_parafac}
\alias{mcr_parafac2}
\alias{mcr_tucker}
\alias{parafac_3way}
\alias{parafac_3wayna}
\alias{parafac_4way}
\alias{parafac_4wayna}
\alias{parafac2_3way}
\alias{parafac2_3wayna}
\alias{parafac2_4way}
\alias{parafac2_4wayna}
\alias{tucker_3way}
\alias{tucker_3wayna}
\alias{tucker_4way}
\alias{tucker_4wayna}

\title{Internal Multi-Way Functions}
\description{
  Internal Multi-Way Functions
}
\details{
  These functions are not to be called by the user.
}
\keyword{ internal }

