\name{multiway-package}
\alias{multiway-package}
\alias{multiway}
\docType{package}
\title{
\packageTitle{multiway}
}
\description{
\packageDescription{multiway}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{multiway}
\packageIndices{multiway}
\code{\link{cpd}} computes the N-way Canonical Polyadic Decomposition of a tensor. \code{\link{indscal}} fits the Individual Differences Scaling model. \code{\link{mcr}} fits the Multiway Covariates Regression model. \code{\link{parafac}} fits the 3-way and 4-way Parallel Factor Analysis-1 model. \code{\link{parafac2}} fits the 3-way and 4-way Parallel Factor Analysis-2 model. \code{\link{sca}} fits the four different Simultaneous Component Analysis models. \code{\link{tucker}} fits the 3-way and 4-way Tucker Factor Analysis model.
}
\author{
\packageAuthor{multiway}

Maintainer: \packageMaintainer{multiway}
}
\references{
Bro, R., & De Jong, S. (1997). A fast non-negativity-constrained least squares algorithm. \emph{Journal of Chemometrics, 11}, 393-401.

Bro, R., & Kiers, H.A.L. (2003). A new efficient method for determining the number of components in PARAFAC models. \emph{Journal of Chemometrics, 17}, 274-286.

Carroll, J. D., & Chang, J-J. (1970). Analysis of individual differences in multidimensional scaling via an n-way generalization of "Eckart-Young" decomposition. \emph{Psychometrika, 35}, 283-319.

Harshman, R. A. (1970). Foundations of the PARAFAC procedure: Models and conditions for an "explanatory" multimodal factor analysis. \emph{UCLA Working Papers in Phonetics, 16}, 1-84.

Harshman, R. A. (1972). PARAFAC2: Mathematical and technical notes. \emph{UCLA Working Papers in Phonetics, 22}, 30-44.

Harshman, R. A., & Lundy, M. E. (1994). PARAFAC: Parallel factor analysis. \emph{Computational Statistics and Data Analysis, 18}, 39-72.

Haughwout, S. P., LaVallee, R. A., & Castle, I-J. P. (2015). 
   Surveillance Report #102: Apparent Per Capita Alcohol Consumption: 
   National, State, and Regional Trends, 1977-2013. Bethesda, MD: NIAAA,
   Alcohol Epidemiologic Data System.

Helwig, N. E. (2013). The special sign indeterminacy of the direct-fitting Parafac2 model: Some implications, cautions, and recommendations, for Simultaneous Component Analysis. \emph{Psychometrika, 78}, 725-739.

Helwig, N. E. (2017). Estimating latent trends in multivariate longitudinal data via Parafac2 with functional and structural constraints. \emph{Biometrical Journal, 59}(4), 783-803.

Helwig, N. E. (in prep). Constrained parallel factor analysis via the R package multiway.

Hitchcock, F. L. (1927). The expression of a tensor or a polyadic as a sum of products. \emph{Journal of Mathematics and Physics, 6}, 164-189.

Kiers, H. A. L., ten Berge, J. M. F., & Bro, R. (1999). PARAFAC2-part I: A direct-fitting algorithm for the PARAFAC2 model. \emph{Journal of Chemometrics, 13}, 275-294.

Kroonenberg, P. M., & de Leeuw, J. (1980). Principal component analysis of three-mode data by means of alternating least squares algorithms. \emph{Psychometrika, 45}, 69-97.

Moore, E. H. (1920). On the reciprocal of the general algebraic matrix. Bulletin of the American Mathematical Society 26, 394-395.

Nephew, T. M., Yi, H., Williams, G. D., Stinson, F. S., & Dufour, M.C., (2004). 
   U.S. Alcohol Epidemiologic Data Reference Manual, Vol. 1, 4th ed. 
   U.S. Apparent Consumption of Alcoholic Beverages Based on State Sales, 
   Taxation, or Receipt Data. Bethesda, MD: NIAAA, Alcohol Epidemiologic Data 
   System. NIH Publication No. 04-5563.

Penrose, R. (1950). A generalized inverse for matrices. Mathematical Proceedings of the Cambridge Philosophical Society 51, 406-413.

Ramsay, J. O. (1988). Monotone regression splines in action. \emph{Statistical Science, 3}, 425-441.

Smilde, A. K., & Kiers, H. A. L. (1999). Multiway covariates regression models. \emph{Journal of Chemometrics, 13}, 31-48.

Timmerman, M. E., & Kiers, H. A. L. (2003). Four simultaneous component models for the analysis of multivariate time series from more than one subject to model intraindividual and interindividual differences. \emph{Psychometrika, 68}, 105-121.

Tucker, L. R. (1966). Some mathematical notes on three-mode factor analysis. \emph{Psychometrika, 31}, 279-311.
}
\keyword{ package }
\seealso{
\code{\link[CMLS:CMLS-package]{CMLS}} ~~
}
\examples{
# See examples for... 
#   cpd (Canonical Polyadic Decomposition)
#   indscal (INividual Differences SCALing)
#   mcr (Multiway Covariates Regression)
#   parafac (Parallel Factor Analysis-1)
#   parafac2 (Parallel Factor Analysis-2)
#   sca (Simultaneous Component Analysis)
#   tucker (Tucker Factor Analysis)
}
